@@ -1,5 +1,9 @@
-const fs = require('fs')
-const path = require('path')
+import fs from 'node:fs'
+import path from 'node:path'
+import { fileURLToPath } from 'node:url';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
 
  const getAppRootDir  = ( ) => {
   let currentDir = __dirname
@@ -9,4 +13,4 @@ const path = require('path')
   return currentDir
 }
 
-exports.getAppRootDir = getAppRootDir
+export {getAppRootDir,  __dirname}
